+++ creche/src/main/java/Dominio/Features/Crianca/ICriancaPostgresRepository.java
@@ -18,9 +18,9 @@ public interface ICriancaPostgresRepository {
 
     Crianca atualizar(Crianca crianca) throws SQLException, Exception;
 
-    Crianca pegar(long id) throws SQLException, Exception;
+    Crianca pegar(int id) throws SQLException, Exception;
 
-    boolean deletar(long id) throws SQLException, Exception;
+    boolean deletar(int id) throws SQLException, Exception;
 
     List<Crianca> pegarTodas() throws SQLException;
 }
